engine/requirements.txt
GitPython
PyYAML
Jinja2
docker
click
pydantic
